feat: add queue assertion helpers and test job for verifying queue states
refs: https://github.com/RonasIT/laravel-helpers/issues/285
