+++ tests/fixtures/TestingTraitTest/queue_states/v11/assert_queue_equals.json
@@ -8,8 +8,8 @@
             ],
             "job": null,
             "connection": null,
-            "queue": null,
-            "delay": null,
+            "queue": "some_queue",
+            "delay": "2020-01-01T00:01:00.000000Z",
             "afterCommit": null,
             "middleware": [],
             "chained": [],
