+++ tests/fixtures/TestingTraitTest/queue_states/v11/assert_queue_equals_as_class_name.json
@@ -6,7 +6,7 @@
             "anotherPayload": [],
             "job": null,
             "connection": null,
-            "queue": null,
+            "queue": "some_queue",
             "delay": null,
             "afterCommit": null,
             "middleware": [],
