+++ tests/fixtures/TestingTraitTest/queue_states/v11/assert_queue_equals_different_jobs.json
@@ -8,7 +8,7 @@
       ],
       "job": null,
       "connection": null,
-      "queue": null,
+      "queue": "some_queue",
       "delay": null,
       "afterCommit": null,
       "middleware": [],
