+++ tests/fixtures/TestingTraitTest/queue_states/v12/assert_queue_equals.json
@@ -8,10 +8,10 @@
             ],
             "job": null,
             "connection": null,
-            "queue": null,
+            "queue": "some_queue",
             "messageGroup": null,
             "deduplicator": null,
-            "delay": null,
+            "delay": "2020-01-01T00:01:00.000000Z",
             "afterCommit": null,
             "middleware": [],
             "chained": [],
