+++ tests/fixtures/TestingTraitTest/queue_states/v12/assert_queue_equals_different_jobs.json
@@ -8,7 +8,7 @@
             ],
             "job": null,
             "connection": null,
-            "queue": null,
+            "queue": "some_queue",
             "messageGroup": null,
             "deduplicator": null,
             "delay": null,
