+++ tests/fixtures/TestingTraitTest/queue_states/v13/assert_queue_equals.json
@@ -8,11 +8,11 @@
             ],
             "job": null,
             "connection": null,
-            "queue": null,
+            "queue": "some_queue",
             "messageGroup": null,
             "deduplicator": null,
             "debounceOwner": "",
-            "delay": null,
+            "delay": "2020-01-01T00:01:00.000000Z",
             "afterCommit": null,
             "middleware": [],
             "chained": [],
