+++ tests/fixtures/TestingTraitTest/queue_states/v13/assert_queue_equals_as_class_name_with_params.json
@@ -8,7 +8,7 @@
             ],
             "job": null,
             "connection": null,
-            "queue": null,
+            "queue": "some_queue",
             "messageGroup": null,
             "deduplicator": null,
             "debounceOwner": "",
